pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trauma-desert"
version = "0.1.0"
description = "Mapping gun violence burden against trauma system capacity in Philadelphia"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Trauma Desert Research Team"}
]
keywords = [
    "trauma",
    "gun violence",
    "public health",
    "geospatial",
    "philadelphia",
    "health equity"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "fiona>=1.9.0",
    "folium>=0.15.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "requests>=2.31.0",
    "openrouteservice>=2.3.3",
    "census>=0.8.21",
    "us>=3.1.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "pytest>=7.4.0",
    "black>=23.12.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/trauma-desert"
Repository = "https://github.com/yourusername/trauma-desert"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | data
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

